connect everything
